pull

+++ application/admin/controller/Upload.php
@@ -367,23 +367,24 @@ class Upload extends Controller
         //获取店铺id
         $store_id=Session::get('store_id');
         //获取小程序二维码
-        // if (file_exists(ROOT_PATH . 'public' . DS . 'uploads'.DS.'D'.$store_id.'.txt')) {
-        //     //检查是否有该文件夹，如果没有就创建，并给予最高权限
-        //     $re=file_get_contents(ROOT_PATH . 'public' . DS . 'uploads'.DS.'D'.$store_id.'.txt');  //小程序二维码
-        // }else{
+        if (file_exists(ROOT_PATH . 'public' . DS . 'uploads'.DS.'D'.$store_id.'.txt')) {
+            //检查是否有该文件夹，如果没有就创建，并给予最高权限
+            $re=file_get_contents(ROOT_PATH . 'public' . DS . 'uploads'.DS.'D'.$store_id.'.txt');  //小程序二维码
+        }else{
             //获取携带参数的小程序的二维码
             $page='pages/logs/logs';
             $qr=new My();
             $qrcode=$qr->mpcode2($page,$store_id);
+            halt($qrcode);
             //把qrcode文件写进文件中，使用的时候拿出来
-            $new_file = ROOT_PATH . 'public' . DS . 'uploads'.DS.'D6.txt';
+            $new_file = ROOT_PATH . 'public' . DS . 'uploads'.DS.'D'.$store_id.'.txt';
                 //检查是否有该文件夹，如果没有就创建，并给予最高权限
                 // mkdir($new_file, 0777);
                 // mkdir($new_file, 750);
             if (file_put_contents($new_file, $qrcode)) {
-                $re=file_get_contents(ROOT_PATH . 'public' . DS . 'uploads'.DS.'D6.txt');
+                $re=file_get_contents(ROOT_PATH . 'public' . DS . 'uploads'.DS.'D'.$store_id.'.txt');
             } 
-        // }
+        }
         //判断是否已授权
         $is_shou=db('miniprogram')->where('store_id',$store_id)->find();
         if($is_shou){
